pyproject.toml
[tool.poetry]
name = "data-analysis-convai"
version = "0.1.0"
description = "Conversation AI for data analysis using Function Calling and Agent Execution. Query database to get result and display it in conversation ux using chainlit"
authors = ["sourabh shrishrimal"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.24.1"
python-dotenv = "^1.0.1"
plotly = "^5.21.0"
psycopg2-binary = "^2.9.9"
chainlit = "^1.0.506"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
